InvalidVariableNameException moved
