+++ web-ui/src/main/java/org/archcnl/domain/common/conceptsandrelations/andtriplets/triplet/Variable.java
@@ -1,7 +1,7 @@
 package org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet;
 
 import java.util.Objects;
-import org.archcnl.domain.input.exceptions.InvalidVariableNameException;
+import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.exceptions.InvalidVariableNameException;
 
 public class Variable extends ObjectType {
